+++ lib/reactrb.rb
@@ -2,18 +2,16 @@
 
 if RUBY_ENGINE == 'opal'
   if `window.React === undefined || window.React.version === undefined`
-    raise "No React.js Available\n\n"\
-          "React.js must be defined before requiring 'reactrb'\n"\
-          "'reactrb' has been tested with react v13, v14, and v15.\n\n"\
-          "IF USING 'react-rails':\n"\
-          "   add 'require \"react\"' immediately before the 'require \"reactive-ruby\" "\
-              "directive in 'views/components.rb'.\n"\
-          "IF USING WEBPACK:\n"\
-          "   add 'react' to your webpack manifest.\n"\
-          "OTHERWISE TO GET THE LATEST TESTED VERSION\n"\
-          "   add 'require \"react-latest\"' immediately before the require of 'reactrb',\n"\
-          "OR TO USE A SPECIFIC VERSION\n"\
-          "   add 'require \"react-v1x\"' immediately before the require of 'reactrb'."
+    raise [
+      "No React.js Available",
+      "",
+      "A global `React` must be defined before requiring 'reactrb'",
+      "",
+      "To USE THE BUILT-IN SOURCE: ",
+      "   add 'require \"react/react-source\"' immediately before the 'require \"reactrb\" directive.",
+      "IF USING WEBPACK:",
+      "   add 'react' to your webpack manifest."
+    ].join("\n")
   end
   require 'react/hash'
   require 'react/top_level'
@@ -50,4 +48,5 @@ else
 
   Opal.append_path File.expand_path('../', __FILE__).untaint
   Opal.append_path File.expand_path('../sources/', __FILE__).untaint
+  require "react/react-source"
 end